+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 +1
 --
 
+  * Spotify dropped support for Safari/WebKit. Issue: tiliado/nuvolaruntime#374
   * The check of WebKitGTK+ version was removed as Nuvola 3.0.9 depends on new enough version (2.16.3).
 
 2.3 - February 12th, 2017